+++ scrapy/contrib/loader/processor.py
@@ -1,7 +1,7 @@
 import warnings
 from scrapy.exceptions import ScrapyDeprecationWarning
 warnings.warn("Module `scrapy.contrib.loader.processor` is deprecated, "
-              "use `scrapy.loader.processor` instead",
+              "use `scrapy.loader.processors` instead",
               ScrapyDeprecationWarning, stacklevel=2)
 
-from scrapy.loader.processor import *
+from scrapy.loader.processors import *
