+++ scrapy/loader/__init__.py
@@ -14,7 +14,7 @@ from scrapy.utils.misc import arg_to_iter, extract_regex
 from scrapy.utils.python import flatten
 
 from .common import wrap_loader_context
-from .processor import Identity
+from .processors import Identity
 
 
 class ItemLoader(object):
