+++ scrapy/loader/processors.py
