+++ tests/test_loader.py
@@ -2,7 +2,7 @@ import unittest
 from functools import partial
 
 from scrapy.loader import ItemLoader
-from scrapy.loader.processor import Join, Identity, TakeFirst, \
+from scrapy.loader.processors import Join, Identity, TakeFirst, \
     Compose, MapCompose, SelectJmes
 from scrapy.item import Item, Field
 from scrapy.selector import Selector
